refactor(web): rename MCP display metadata context
