+++ packages/web/src/ee/features/chat/components/chatThread/tools/mcpToolComponent.test.tsx
@@ -1,7 +1,7 @@
 import { render, screen } from '@testing-library/react';
 import type { DynamicToolUIPart } from 'ai';
 import { describe, expect, test } from 'vitest';
-import { McpToolNameContext } from '@/ee/features/chat/mcpServerIconContext';
+import { McpToolNameContext } from '@/ee/features/chat/mcpDisplayMetadataContext';
 import { getMcpToolDisplayParts, McpToolComponent } from './mcpToolComponent';
 
 describe('getMcpToolDisplayParts', () => {
