+++ packages/web/src/ee/features/chat/mcpDisplayMetadataContext.tsx
